+++ toolkit/scaffold/project/notes/cli.py
@@ -6,6 +6,7 @@ import click
 
 from toolkit.scaffold.project.command import generate_create_project_command
 from toolkit.scaffold.project.notes.article import create_article
+from toolkit.scaffold.project.notes.image import offline_images
 from toolkit.scaffold.project.notes.toc import extract_toc_from_folder
 from toolkit.scaffold.project.template import TEMPLATE_NOTES_PATH
 
@@ -35,6 +36,8 @@ create_notes_project.add_command(create_article, "article")
 create_notes_project.add_command(create_all, "all")
 
 create_notes_project.add_command(extract_toc_from_folder, "toc")
+
+create_notes_project.add_command(offline_images, "image")
 
 if __name__ == "__main__":
     sys.exit(create_notes_project())  # pragma: no cover